#18 refactor(logbuffer, setupinfo): moved to internal pkg dir
