+++ k2s/cmd/k2s/addons/status/print_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/cmd/common"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	r "github.com/siemens-healthineers/k2s/internal/reflection"
