+++ k2s/cmd/k2s/cmd/addons/addons.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils/psexecutor"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	"path/filepath"
