+++ k2s/cmd/k2s/cmd/addons/cmd/export/export.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	p "github.com/siemens-healthineers/k2s/cmd/k2s/cmd/params"
