+++ k2s/cmd/k2s/cmd/common/common.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils/logging"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	"github.com/pterm/pterm"
 )
