+++ k2s/cmd/k2s/cmd/common/common_test.go
@@ -7,7 +7,7 @@ import (
 	"log/slog"
 	"testing"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	"github.com/go-logr/logr"
 	. "github.com/onsi/ginkgo/v2"
