+++ k2s/cmd/k2s/cmd/image/export.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	p "github.com/siemens-healthineers/k2s/cmd/k2s/cmd/params"
