+++ k2s/cmd/k2s/cmd/image/registry/list.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/cmd/common"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
