+++ k2s/cmd/k2s/cmd/image/registry/switch.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
