+++ k2s/cmd/k2s/cmd/install/core/installer.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils/psexecutor"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	"github.com/siemens-healthineers/k2s/internal/version"
