+++ k2s/cmd/k2s/cmd/start/start.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 )
 
 var Startk8sCmd = &cobra.Command{
