+++ k2s/cmd/k2s/cmd/status/load/statusLoader.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 )
 
 type StatusLoader struct {
