+++ k2s/cmd/k2s/cmd/system/scp/scp.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	"github.com/spf13/cobra"
 )
