+++ k2s/cmd/k2s/cmd/system/upgrade/upgrade.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 )
 
 var upgradeCommandShortDescription = "Upgrade your current cluster to this version"
