+++ k2s/cmd/k2s/cmd/uninstall/uninstall.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/utils"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 )
 
 var (
