+++ k2s/cmd/k2s/config/config.go
@@ -8,7 +8,7 @@ import (
 
 	d "github.com/siemens-healthineers/k2s/cmd/k2s/config/defs"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 )
 
 type ConfigLoader interface {
