+++ k2s/cmd/k2s/config/defs/defs.go
@@ -4,7 +4,7 @@
 package defs
 
 import (
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 )
 
 type SmbHostType string
