+++ k2s/internal/setupinfo/setupinfo.go
