+++ k2s/test/e2e/cli/cmd/image/nosetup/image_no-setup_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/cmd/image"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
