+++ k2s/test/e2e/cli/cmd/status/setuprequired/systemstopped/status_system-stopped_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/siemens-healthineers/k2s/cmd/k2s/cmd/status/load"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	"testing"
