+++ k2s/test/framework/k2s/kubeproxy.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"path/filepath"
 
-	"github.com/siemens-healthineers/k2s/cmd/k2s/setupinfo"
+	"github.com/siemens-healthineers/k2s/internal/setupinfo"
 
 	//lint:ignore ST1001 test framework code
 	. "github.com/onsi/ginkgo/v2"
